requirements.txt
# Python dependencies for AI Upscaling
numpy
opencv-python
Pillow
torch
torchvision
realesrgan
scikit-image

